[Silver III] Title: 주유소, Time: 84 ms, Memory: 2800 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -8,11 +8,11 @@
 #define MAX 100001
 using namespace std;
 
-long long roadLength[MAX];
-long long price[MAX];
+int roadLength[MAX];
+int price[MAX];
 int main()
 {
-	long long n;
+	int n;
 	cin >> n;
 	for (int i = 0; i < n - 1; i++)
 	{
@@ -23,13 +23,13 @@ int main()
 	{
 		cin >> price[i];
 	}
-	long long minValue = INT_MAX;
-	int result = 0;
+	int minValue = INT_MAX;
+	long long result = 0;
 	for (int i = 0; i < n; i++)
 	{
 		minValue = min(minValue, price[i]);
 
-		result += minValue * roadLength[i];
+		result += (long long)minValue * roadLength[i];
 
 	}
